Move ContactGroupForm to forms.py file
