+++ tracpro/groups/views.py
@@ -2,31 +2,13 @@ from __future__ import absolute_import, unicode_literals
 
 from dash.orgs.views import OrgPermsMixin
 
-from django import forms
 from django.http import HttpResponseRedirect, JsonResponse
 from django.utils.translation import ugettext_lazy as _
 
 from smartmin.users.views import SmartCRUDL, SmartListView, SmartFormView
 
 from .models import Group, Region
-
-
-class ContactGroupsForm(forms.Form):
-    groups = forms.MultipleChoiceField(choices=(), label=_("Groups"),
-                                       help_text=_("Contact groups to use."))
-
-    def __init__(self, *args, **kwargs):
-        org = kwargs.pop('org')
-        initial = kwargs.pop('initial')
-
-        super(ContactGroupsForm, self).__init__(*args, **kwargs)
-
-        choices = []
-        for group in org.get_temba_client().get_groups():
-            choices.append((group.uuid, "%s (%d)" % (group.name, group.size)))
-
-        self.fields['groups'].choices = choices
-        self.fields['groups'].initial = initial
+from .forms import ContactGroupsForm
 
 
 class RegionCRUDL(SmartCRUDL):
